refactor(runtime): W6 — relocate bootstrap SPI/egress to Doc 2 homes
bootstrap.AbstractRuntimeAgentHandler -> engine.spi.AbstractAgentRuntimeHandler; bootstrap.AccessNotificationClient -> access.output.EngineOutputSink (Doc 2 §1.2). Added missing import in AgentServiceBootstrapConfiguration (was same-package). Pure move+rename, behavior preserved.

NOTE: the pure-Java app.RuntimeApp/RuntimeHost/LocalA2aRuntimeHost entry (Doc 2 §10) and full bootstrap retirement are NEW bootable behavioral code that cannot be runtime-verified in this env (no LLM/Mockito; example boot path self-skips) — archived with spec in the plan file rather than shipped unverified.

Verified: clean test-compile + non-Mockito tests green; example e2e green.
